# docs/swagger.yaml
definitions:
  models.CommentInput:
    properties:
      content:
        type: string
      wish_id:
        type: integer
    required:
    - content
    - wish_id
    type: object
  models.WishInput:
    properties:
      category:
        type: string
      content:
        type: string
      title:
        type: string
    required:
    - category
    - content
    - title
    type: object
info:
  contact: {}
paths:
  /api/v1/comments:
    get:
      consumes:
      - application/json
      description: 특정 소원에 대한 댓글을 조회합니다.
      parameters:
      - description: Wish ID
        in: query
        name: wish_id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of comments per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 댓글 조회
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: 소원에 댓글을 작성합니다.
      parameters:
      - description: ID와 내용을 입력하세요.
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.CommentInput'
      produces:
      - application/json
      responses: {}
      summary: 댓글 작성
      tags:
      - Comments
  /api/v1/comments/{id}:
    delete:
      consumes:
      - application/json
      description: ID로 특정 댓글을 삭제합니다.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 댓글 삭제
      tags:
      - Comments
  /api/v1/wishes:
    get:
      consumes:
      - application/json
      description: 조회하려는 소원의 상태와 페이지(번호, 크기)를 선택할 수 있습니다.
      parameters:
      - description: '승인: approved, 보류: pending, 거절: rejected'
        in: query
        name: status
        type: string
      - default: 1
        description: 페이지 번호
        in: query
        name: page
        type: integer
      - default: 10
        description: 페이지 크기
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 소원 목록 조회
      tags:
      - Wishes
  /api/v1/wishes/:
    post:
      consumes:
      - application/json
      description: 제목, 내용, 카테고리를 입력하여 소원을 등록해보세요.
      parameters:
      - description: 카테고리의 종류는 진로, 건강, 인간 관계, 돈, 목표, 학업/성적, 기타 총 7가지입니다.
        in: body
        name: wish
        required: true
        schema:
          $ref: '#/definitions/models.WishInput'
      produces:
      - application/json
      responses: {}
      summary: 소원 등록
      tags:
      - Wishes
  /api/v1/wishes/{id}:
    delete:
      consumes:
      - application/json
      description: ID를 입력하여 해당 소원을 삭제할 수 있습니다.
      parameters:
      - description: Wish ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 소원 삭제
      tags:
      - Wishes
    get:
      consumes:
      - application/json
      description: ID를 입력하여 승인된 소원을 조회할 수 있습니다.
      parameters:
      - description: Wish ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 개별 소원 조회
      tags:
      - Wishes
  /api/v1/wishes/{id}/status:
    put:
      consumes:
      - application/json
      description: ID를 입력하여 해당 소원의 상태를 변경할 수 있습니다.
      parameters:
      - description: Wish ID
        in: path
        name: id
        required: true
        type: integer
      - description: '승인: approved, 거절: rejected'
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 개별 소원 상태 변경
      tags:
      - Wishes
  /api/v1/wishes/status:
    put:
      consumes:
      - application/json
      description: 모든 소원을 승인 또는 거절 상태로 변경할 수 있습니다.
      parameters:
      - description: '승인: approved, 거절: rejected'
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 모든 소원 상태 변경
      tags:
      - Wishes
swagger: "2.0"